Move AddUserRole to Panel
