+++ src/components/Roles/RolesByUser.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { useRoleManagement, useAllUserRolesByUser, SPRole } from "api/RolesApi";
 import { PeoplePicker } from "components/PeoplePicker/PeoplePicker";
 import { IPerson } from "api/UserApi";
@@ -13,6 +13,8 @@ import {
 } from "@fluentui/react";
 import { ContactIcon } from "@fluentui/react-icons-mdl2";
 import { Label, makeStyles, Text } from "@fluentui/react-components";
+import { AddUserRolePanel } from "components/Roles/AddUserRolePanel";
+import { useBoolean } from "@fluentui/react-hooks";
 
 /* FluentUI Styling */
 const useStyles = makeStyles({
@@ -37,22 +39,14 @@ const useStyles = makeStyles({
 export const RolesByUser: React.FunctionComponent = () => {
   const classes = useStyles();
   const { data: allRolesByUser } = useAllUserRolesByUser();
-  const [items, setItems] = useState<SPRole[]>([]);
   const [selectedUser, setSelectedUser] = useState<IPerson>();
   const [selection] = useState(new Selection());
 
-  useEffect(() => {
-    if (typeof selectedUser?.EMail === "string") {
-      const newItems = allRolesByUser?.get(selectedUser.EMail);
-      if (newItems === undefined) {
-        setItems([]);
-      } else {
-        setItems(newItems);
-      }
-    } else {
-      setItems([]);
-    }
-  }, [allRolesByUser, selectedUser?.EMail]);
+  /* Boolean state for determining whether or not the AddUserRolePanel is shown */
+  const [isAddPanelOpen, { setTrue: showAddPanel, setFalse: hideAddPanel }] =
+    useBoolean(false);
+
+  const items = allRolesByUser?.get(selectedUser?.EMail || "") || [];
 
   const columns: IColumn[] = [
     {
@@ -80,38 +74,52 @@ export const RolesByUser: React.FunctionComponent = () => {
         }
       },
     },
+    {
+      key: "add",
+      text: "Add Role",
+      iconProps: { iconName: "Add" },
+      onClick: showAddPanel,
+    },
   ];
 
   return (
-    <div className={classes.formContainer}>
-      <div className={classes.fieldContainer}>
-        <Label size="small" weight="semibold" className={classes.fieldLabel}>
-          <ContactIcon className={classes.fieldIcon} />
-          User
-        </Label>
-        <PeoplePicker
-          ariaLabel={"Select the user to view the roles of"}
-          updatePeople={(selectedUser) => {
-            if (selectedUser) {
-              setSelectedUser(selectedUser[0]);
-            }
-          }}
-          selectedItems={selectedUser ? selectedUser : []}
-        ></PeoplePicker>
+    <>
+      <div className={classes.formContainer}>
+        <div className={classes.fieldContainer}>
+          <Label size="small" weight="semibold" className={classes.fieldLabel}>
+            <ContactIcon className={classes.fieldIcon} />
+            User
+          </Label>
+          <PeoplePicker
+            ariaLabel={"Select the user to view the roles of"}
+            updatePeople={(selectedUser) => {
+              if (selectedUser) {
+                setSelectedUser(selectedUser[0]);
+              }
+            }}
+            selectedItems={selectedUser ? selectedUser : []}
+          ></PeoplePicker>
+        </div>
+        <div className={classes.fieldContainer}>
+          <CommandBar items={commandItems}></CommandBar>
+          <DetailsList
+            items={items}
+            columns={columns}
+            selectionMode={SelectionMode.multiple}
+            constrainMode={ConstrainMode.unconstrained}
+            selection={selection}
+          ></DetailsList>
+          {items.length === 0 && (
+            <Text>No roles assigned to the selected user.</Text>
+          )}
+        </div>
       </div>
-      <div className={classes.fieldContainer}>
-        <CommandBar items={commandItems}></CommandBar>
-        <DetailsList
-          items={items}
-          columns={columns}
-          selectionMode={SelectionMode.multiple}
-          constrainMode={ConstrainMode.unconstrained}
-          selection={selection}
-        ></DetailsList>
-        {items.length === 0 && (
-          <Text>No roles assigned to the selected user.</Text>
-        )}
-      </div>
-    </div>
+      <AddUserRolePanel
+        isAddPanelOpen={isAddPanelOpen}
+        onAddCancel={hideAddPanel}
+        onAdd={hideAddPanel}
+        defaultUser={selectedUser}
+      />
+    </>
   );
 };
